Can now see group classes when registering Made it so admins dont make PTS, members do

+++ src/main/java/Member.java
@@ -9,7 +9,7 @@ public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_id")
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String name;
@@ -43,7 +43,7 @@ public class Member {
         this.dateOfBirth = dateOfBirth;
     }
 
-    public Long getId() {return id;}
+    public int getId() {return id;}
 
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
